// apps/frontend/src/gen/example_pb.d.ts
// @generated by protoc-gen-es v2.10.1
// @generated from file example.proto (package datifyy.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file example.proto.
 */
export declare const file_example: GenFile;

/**
 * @generated from message datifyy.v1.User
 */
export declare type User = Message<"datifyy.v1.User"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: string email = 3;
   */
  email: string;
};

/**
 * Describes the message datifyy.v1.User.
 * Use `create(UserSchema)` to create a new message.
 */
export declare const UserSchema: GenMessage<User>;

/**
 * @generated from message datifyy.v1.GetUserRequest
 */
export declare type GetUserRequest = Message<"datifyy.v1.GetUserRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string yt = 2;
   */
  yt: string;
};

/**
 * Describes the message datifyy.v1.GetUserRequest.
 * Use `create(GetUserRequestSchema)` to create a new message.
 */
export declare const GetUserRequestSchema: GenMessage<GetUserRequest>;

/**
 * @generated from message datifyy.v1.GetUserResponse
 */
export declare type GetUserResponse = Message<"datifyy.v1.GetUserResponse"> & {
  /**
   * @generated from field: datifyy.v1.User user = 1;
   */
  user?: User;
};

/**
 * Describes the message datifyy.v1.GetUserResponse.
 * Use `create(GetUserResponseSchema)` to create a new message.
 */
export declare const GetUserResponseSchema: GenMessage<GetUserResponse>;

/**
 * @generated from service datifyy.v1.UserService
 */
export declare const UserService: GenService<{
  /**
   * @generated from rpc datifyy.v1.UserService.GetUser
   */
  getUser: {
    methodKind: "unary";
    input: typeof GetUserRequestSchema;
    output: typeof GetUserResponseSchema;
  },
}>;